updated for version 7.3.1074 Problem:    Compiler warning for printf format. (Manuel Ortega) Solution:   Add type casts.

+++ src/if_py_both.h
@@ -3807,7 +3807,7 @@ RangeRepr(RangeObject *self)
 	    name = "";
 
 	return PyString_FromFormat("<range %s (%d:%d)>",
-				    name, self->start, self->end);
+				    name, (int)self->start, (int)self->end);
     }
 }
